Domain change from BudgetParser to BudgetFolder
The BudgetParser class wasn't really parsing any budget data. It was,
instead just identifying the budget file by parsing several metadata
files in the YNAB budget directory. This change reflects the domain as
being the YNAB Budget Folder and provides methods for parsing the
metadata.

I'm still fairly unhappy with the Ynab::BudgetFolder#devices test, as it
feels very brittle and relies heavily upon the implementation of the
'#devices' method that it's testing. I'll look for ways to improve that.